requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.23.2
python-dotenv>=1.0.0
python-multipart>=0.0.6
pydantic>=2.8.0
requests>=2.31.0

# LlamaIndex and embeddings
llama-index-core>=0.12.0
llama-index-embeddings-openai>=0.3.1
llama-index-llms-openai>=0.3.1
llama-index-readers-file>=0.1.6

# LiveKit integration
livekit>=1.0.8
livekit-agents>=1.0.13
livekit-plugins-deepgram>=1.0.13
livekit-plugins-openai>=1.0.13
livekit-plugins-cartesia>=1.0.13
livekit-plugins-silero>=1.0.13

# OpenAI
openai>=1.1.0

# Audio processing (for voice agents)
pyaudio>=0.2.14
soundfile>=0.13.1

# Other utilities
tqdm>=4.66.1
loguru>=0.7.0 